fragment editor base class, reviews addressed
