+++ studio/src/editor-panel.js
@@ -1,0 +1,279 @@
+import { html, css, nothing } from 'lit';
+import { FragmentEditorBase } from './fragment-editor-base.js';
+
+class EditorPanel extends FragmentEditorBase {
+    static properties = {
+        showToast: { type: Function },
+        fragment: { type: Object, attribute: false },
+        source: { type: Object },
+        bucket: { type: String },
+        disabled: { type: Boolean },
+        hasChanges: { type: Boolean },
+    };
+
+    static styles = css`
+        sp-divider {
+            margin: 16px 0;
+        }
+
+        merch-card-editor {
+            display: contents;
+        }
+
+        #actions {
+            display: flex;
+            justify-content: end;
+        }
+
+        sp-textfield {
+            width: 360px;
+        }
+    `;
+
+    constructor() {
+        super();
+        this.disabled = false;
+        this.close = this.close.bind(this);
+        this.handleClose = this.handleClose.bind(this);
+        this.handleKeyDown = this.handleKeyDown.bind(this);
+    }
+
+    connectedCallback() {
+        super.connectedCallback();
+        document.addEventListener('keydown', this.handleKeyDown);
+    }
+
+    disconnectedCallback() {
+        super.disconnectedCallback();
+        document.removeEventListener('keydown', this.handleKeyDown);
+    }
+
+    handleKeyDown(event) {
+        if (event.code === 'Escape') this.close();
+        if (event.code === 'ArrowLeft' && event.shiftKey)
+            this.updatePosition('left');
+        if (event.code === 'ArrowRight' && event.shiftKey)
+            this.updatePosition('right');
+    }
+
+    handleClose(e) {
+        if (e.target === this) return;
+        e.stopPropagation();
+    }
+
+    openFragmentInOdin() {
+        const parent = this.fragment?.path.split('/').slice(0, -1).join('/');
+        window.open(
+            `https://experience.adobe.com/?repo=author-p22655-${this.bucket}.adobeaemcloud.com#/@odin02/aem/cf/admin${parent}?appId=aem-cf-admin&q=${this.fragment?.fragmentName}`,
+            '_blank',
+        );
+    }
+
+    async copyToUse() {
+        //@TODO make it generic.
+        const code = `<merch-card><aem-fragment fragment="${this.fragment?.id}"></aem-fragment></merch-card>`;
+        try {
+            await navigator.clipboard.writeText(code);
+            this.showToast('Code copied to clipboard', 'positive');
+        } catch (e) {
+            this.showToast('Failed to copy code to clipboard', 'negative');
+        }
+    }
+
+    #updateFragmentInternal(event) {
+        const fieldName = event.target.dataset.field;
+        let value = event.target.value;
+        this.fragmentStore.updateFieldInternal(fieldName, value);
+        this.hasChanges = true;
+    }
+
+    get fragmentEditorToolbar() {
+        return html`<div id="editor-toolbar">
+            <sp-action-group
+                aria-label="Fragment actions"
+                role="group"
+                size="m"
+                compact
+                emphasized
+                quiet
+            >
+                <sp-action-button
+                    label="Move left"
+                    title="Move left"
+                    value="left"
+                    id="move-left"
+                    @click="${() => this.updatePosition('left')}"
+                >
+                    <sp-icon-chevron-left slot="icon"></sp-icon-chevron-left>
+                    <sp-tooltip self-managed placement="bottom"
+                        >Move left</sp-tooltip
+                    >
+                </sp-action-button>
+                <sp-action-button
+                    label="Save"
+                    title="Save changes"
+                    value="save"
+                    @click="${this.aemAction(
+                        this.repository.saveFragment,
+                        true,
+                    )}"
+                    ?disabled=${this.disabled}
+                >
+                    <sp-icon-save-floppy slot="icon"></sp-icon-save-floppy>
+                    <sp-tooltip self-managed placement="bottom"
+                        >Save changes</sp-tooltip
+                    >
+                </sp-action-button>
+                <sp-action-button
+                    label="Discard"
+                    title="Discard changes"
+                    value="discard"
+                    @click="${this.discardChanges}"
+                    ?disabled=${this.disabled || !this.hasChanges}
+                >
+                    <sp-icon-undo slot="icon"></sp-icon-undo>
+                    <sp-tooltip self-managed placement="bottom"
+                        >Discard changes</sp-tooltip
+                    >
+                </sp-action-button>
+                <sp-action-button
+                    label="Clone"
+                    value="clone"
+                    @click="${this.aemAction(this.repository.copyFragment)}"
+                    ?disabled=${this.disabled}
+                >
+                    <sp-icon-duplicate slot="icon"></sp-icon-duplicate>
+                    <sp-tooltip self-managed placement="bottom"
+                        >Clone</sp-tooltip
+                    >
+                </sp-action-button>
+                <sp-action-button
+                    label="Publish"
+                    value="publish"
+                    @click="${this.aemAction(this.repository.publishFragment)}"
+                    ?disabled=${this.disabled}
+                >
+                    <sp-icon-publish-check slot="icon"></sp-icon-publish-check>
+                    <sp-tooltip self-managed placement="bottom"
+                        >Publish</sp-tooltip
+                    >
+                </sp-action-button>
+                <sp-action-button
+                    label="Unpublish"
+                    value="unpublish"
+                    @click="${this.aemAction(
+                        this.repository.unpublishFragment,
+                    )}"
+                    disabled
+                >
+                    <sp-icon-publish-remove
+                        slot="icon"
+                    ></sp-icon-publish-remove>
+                    <sp-tooltip self-managed placement="bottom"
+                        >Unpublish</sp-tooltip
+                    >
+                </sp-action-button>
+                <sp-action-button
+                    label="Open in Odin"
+                    value="open"
+                    @click="${this.openFragmentInOdin}"
+                    ?disabled=${this.disabled}
+                >
+                    <sp-icon-open-in slot="icon"></sp-icon-open-in>
+                    <sp-tooltip self-managed placement="bottom"
+                        >Open in Odin</sp-tooltip
+                    >
+                </sp-action-button>
+                <sp-action-button
+                    label="Use"
+                    value="use"
+                    @click="${this.copyToUse}"
+                    ?disabled=${this.disabled}
+                >
+                    <sp-icon-code slot="icon"></sp-icon-code>
+                    <sp-tooltip self-managed placement="bottom">Use</sp-tooltip>
+                </sp-action-button>
+                <sp-action-button
+                    label="Delete fragment"
+                    value="delete"
+                    @click="${this.aemAction(this.repository.deleteFragment)}"
+                    ?disabled=${this.disabled}
+                >
+                    <sp-icon-delete-outline
+                        slot="icon"
+                    ></sp-icon-delete-outline>
+                    <sp-tooltip self-managed placement="bottom"
+                        >Delete fragment</sp-tooltip
+                    >
+                </sp-action-button>
+                <sp-action-button
+                    title="Close"
+                    label="Close"
+                    value="close"
+                    @click="${this.close}"
+                    ?disabled=${this.disabled}
+                >
+                    <sp-icon-close-circle slot="icon"></sp-icon-close-circle>
+                    <sp-tooltip self-managed placement="bottom"
+                        >Close</sp-tooltip
+                    >
+                </sp-action-button>
+                <sp-action-button
+                    label="Move right"
+                    title="Move right"
+                    value="right"
+                    id="move-right"
+                    @click="${() => this.updatePosition('right')}"
+                >
+                    <sp-icon-chevron-right slot="icon"></sp-icon-chevron-right>
+                    <sp-tooltip self-managed placement="bottom"
+                        >Move right</sp-tooltip
+                    >
+                </sp-action-button>
+            </sp-action-group>
+        </div>`;
+    }
+
+    get fragmentEditor() {
+        return html`
+            ${this.fragment
+                ? html`
+                      ${this.fragmentEditorToolbar}
+                      <p>${this.fragment.path}</p>
+                      <p>Fragment details (not shown on the card)</p>
+                      <sp-field-label for="fragment-title"
+                          >Fragment Title</sp-field-label
+                      >
+                      <sp-textfield
+                          placeholder="Enter fragment title"
+                          id="fragment-title"
+                          data-field="title"
+                          value="${this.fragment.title}"
+                          @input=${this.#updateFragmentInternal}
+                          ?disabled=${this.disabled}
+                      ></sp-textfield>
+                      <sp-field-label for="fragment-description"
+                          >Fragment Description</sp-field-label
+                      >
+                      <sp-textfield
+                          placeholder="Enter fragment description"
+                          id="fragment-description"
+                          data-field="description"
+                          multiline
+                          value="${this.fragment.description}"
+                          @input=${this.#updateFragmentInternal}
+                          ?disabled=${this.disabled}
+                      ></sp-textfield>
+                      <sp-divider size="s"></sp-divider>
+                  `
+                : nothing}
+        `;
+    }
+
+    render() {
+        if (!this.fragment) return nothing;
+        return html`${this.fragmentEditor} ${this.selectFragmentDialog}`;
+    }
+}
+
+customElements.define('editor-panel', EditorPanel);
