+++ studio/style.css
@@ -137,7 +137,7 @@ sp-underlay + sp-dialog {
     left: var(--editor-left);
     right: var(--editor-right);
     height: 100vh;
-    width: 440px;
+    width: 460px;
     background-color: var(--spectrum-white);
     padding: 20px;
     overflow-y: auto;
